// src/lib/decorator/decorator.schema.d.ts
import { Path } from '@angular-devkit/core';

export interface DecoratorOptions {
  /**
   * The name of the decorator.
   */
  name: string;
  /**
   * The path to create the decorator.
   */
  path?: string | Path;
  /**
   * The source root path
   */
  sourceRoot?: string;
  /**
   * Flag to indicate if a directory is created.
   */
  flat?: boolean;
}